requirements.txt
cbor2==5.8.0
colorzero==2.0
evdev==1.9.2
gpiozero==2.0.1
lgpio==0.2.2.0
luma.core==2.5.3
luma.oled==3.14.0
pillow==12.1.0
RPi.GPIO==0.7.1
rpi-lgpio==0.6
smbus2==0.6.0